Improve the search for a local IPv4 address
The demo client scans all local network interfaces until it finds a
non-loopback IPv4 address suitable for communicating with the peer,
which is expected to also run on this host.

This change adds better logging, exclusion of Docker docker*/veth*
interfaces, and allows the chosen local interface to be overridden on
the command line.